feat(Select): Replace deprecated implementation
