+++ packages/react-component-library/src/components/Select/ClearButton.tsx
