+++ packages/react-component-library/src/components/Select/Select.test.tsx
@@ -1,363 +1,365 @@
-import React, { useState } from 'react'
+import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
-import {
-  fireEvent,
-  render,
-  RenderResult,
-  waitFor,
-} from '@testing-library/react'
-import { IconAnchor } from '@defencedigital/icon-library'
+import { fireEvent, render, RenderResult } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
+import { ColorAction000, ColorDanger800 } from '@defencedigital/design-tokens'
+import {
+  IconAgriculture,
+  IconAnchor,
+  IconRemove,
+} from '@defencedigital/icon-library'
 
+import { BORDER_WIDTH } from '../../styled-components'
+import { COMPONENT_SIZE } from '../Forms'
 import { Select } from '.'
-import { Button } from '../Button'
-
-const options = [
-  {
-    badge: 'Badge 1',
-    label: 'Option 1',
-    value: '1',
-  },
-  {
-    badge: 'Badge 2',
-    label: 'Option 2',
-    value: '2',
-  },
-]
+import { SelectOption } from './SelectOption'
 
 describe('Select', () => {
+  let onChangeSpy: (value: string | null) => void
   let wrapper: RenderResult
-  let onChangeSpy: (event: any) => void
 
-  describe('when provided with all props', () => {
+  describe('when using the default prop values', () => {
     beforeEach(() => {
+      const canShowFourthOption = false
+
       onChangeSpy = jest.fn()
 
       wrapper = render(
-        <Select
-          className="rn-select--modifier"
-          label="Select label"
-          name="select-name"
-          onChange={onChangeSpy}
-          options={options}
-          value="2"
-        />
+        <Select id="select-id" label="Label" onChange={onChangeSpy}>
+          <SelectOption value="one">One</SelectOption>
+          <SelectOption value="two">Two</SelectOption>
+          <SelectOption value="three">Three</SelectOption>
+          {canShowFourthOption && (
+            <SelectOption value="three">Three</SelectOption>
+          )}
+        </Select>
       )
     })
 
-    it('should add the CSS modifier', () => {
-      expect(wrapper.container.children[0].classList).toContain(
-        'rn-select--modifier'
+    it('sets the `aria-labelledby` attributes', () => {
+      const labelId = wrapper.getByTestId('select-label').getAttribute('id')
+      expect(wrapper.getByTestId('select')).toHaveAttribute(
+        'aria-labelledby',
+        labelId
+      )
+      expect(wrapper.getByTestId('select-input')).toHaveAttribute(
+        'aria-labelledby',
+        labelId
       )
     })
 
-    it('should render the label', () => {
-      expect(wrapper.getByText('Select label')).toBeInTheDocument()
+    it('sets the custom `id` of the input', () => {
+      expect(wrapper.getByTestId('select-input')).toHaveAttribute(
+        'id',
+        'select-id'
+      )
     })
 
-    it('should render the select with the selected value', () => {
-      expect(
-        wrapper.getByTestId('select-single-value-label')
-      ).toHaveTextContent('Option 2')
+    it('renders the label', () => {
+      expect(wrapper.getByText('Label')).toBeInTheDocument()
     })
 
-    it('should add the name', () => {
-      const input = wrapper.getByTestId('react-select-vendor-input')
+    it('sets the `htmlFor` of the label', () => {
+      const expected = wrapper.getByTestId('select-input').getAttribute('id')
 
-      expect(input.getAttribute('name')).toEqual('select-name')
+      expect(wrapper.getByTestId('select-label')).toHaveAttribute(
+        'for',
+        expected
+      )
     })
 
-    describe('when the select is clicked', () => {
+    it('does not set the value', () => {
+      expect(wrapper.getByTestId('select-input')).toHaveValue('')
+    })
+
+    it('does not display a clear button', () => {
+      expect(wrapper.queryAllByTestId('select-clear-button')).toHaveLength(0)
+    })
+
+    it('does not display the items', () => {
+      expect(wrapper.queryAllByTestId('select-option')).toHaveLength(0)
+    })
+
+    it.skip('does not show the arrow button in a hover state', () => {
+      expect(wrapper.getByTestId('select-arrow-button')).toHaveStyleRule(
+        'background-color',
+        'transparent'
+      )
+    })
+
+    describe('when the mouse enters the input', () => {
       beforeEach(() => {
-        const input = wrapper.getByTestId('react-select-vendor-input')
-        fireEvent.focus(input)
-        fireEvent.keyDown(input, {
-          key: 'ArrowDown',
-          code: 40,
-        })
+        fireEvent.mouseEnter(wrapper.getByTestId('select-input'))
       })
 
-      it('should render two options', () => {
-        expect(wrapper.getAllByTestId('select-option-label')).toHaveLength(2)
+      it.skip('shows the arrow button in a hover state', () => {
+        expect(wrapper.getByTestId('select-arrow-button')).toHaveStyleRule(
+          'background-color',
+          ColorAction000
+        )
       })
 
-      describe('when the first option is clicked', () => {
+      describe('and the mouse leaves the input', () => {
         beforeEach(() => {
-          wrapper.getByText('Option 1').click()
+          fireEvent.mouseLeave(wrapper.getByTestId('select-input'))
         })
 
-        it('should render a clear button', () => {
-          const control = wrapper.container.children[0].children[1]
-          const indicators = control.children[1]
-          const firstIndicator = indicators.children[0]
-
-          expect(firstIndicator.classList).toContain(
-            'rn-select__clear-indicator'
+        it.skip('shows the arrow button in a hover state', () => {
+          expect(wrapper.getByTestId('select-arrow-button')).toHaveStyleRule(
+            'background-color',
+            'transparent'
           )
         })
+      })
+    })
 
-        it('should render the select with the selected value', () => {
-          return waitFor(() => {
-            expect(
-              wrapper.getByTestId('select-single-value-label')
-            ).toHaveTextContent('Option 1')
-          })
+    describe('when the arrow button is clicked', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByTestId('select-arrow-button'))
+      })
+
+      it('displays the items', () => {
+        const options = wrapper.getAllByTestId('select-option')
+
+        expect(options[0]).toHaveTextContent('One')
+        expect(options[1]).toHaveTextContent('Two')
+        expect(options[2]).toHaveTextContent('Three')
+        expect(options).toHaveLength(3)
+      })
+
+      describe('and the second item is clicked', () => {
+        beforeEach(() => {
+          userEvent.click(wrapper.getByText('Two'))
         })
 
-        it('should call the `onChange` spy once', () => {
+        it('hides the items', () => {
+          expect(wrapper.queryAllByTestId('select-option')).toHaveLength(0)
+        })
+
+        it('sets the value', () => {
+          expect(wrapper.getByTestId('select-input')).toHaveValue('Two')
+        })
+
+        it('calls the `onChange` callback', () => {
           expect(onChangeSpy).toHaveBeenCalledTimes(1)
-          expect(onChangeSpy).toHaveBeenCalledWith({
-            target: { name: 'select-name', value: '1' },
+          expect(onChangeSpy).toHaveBeenCalledWith('two')
+        })
+
+        describe('and the clear button is clicked', () => {
+          beforeEach(() => {
+            userEvent.click(wrapper.getByTestId('select-clear-button'))
+          })
+
+          it('resets the value', () => {
+            expect(wrapper.getByTestId('select-input')).toHaveValue('')
+          })
+
+          it('calls the `onChange` callback', () => {
+            expect(onChangeSpy).toHaveBeenCalledWith(null)
           })
         })
       })
     })
-  })
 
-  describe('when a downstream consumer provides a data-testid', () => {
-    beforeEach(() => {
-      wrapper = render(<Select options={options} data-testid="select-1" />)
-    })
-
-    it('should not find the input using the default `data-testid`', () => {
-      expect(
-        wrapper.queryAllByTestId('react-select-vendor-input')
-      ).toHaveLength(0)
-    })
-
-    it('should find the input using the new `data-testid`', () => {
-      expect(wrapper.getByTestId('select-1')).toBeInTheDocument()
-    })
-  })
-
-  describe('when provided with options that have icons', () => {
-    beforeEach(() => {
-      onChangeSpy = jest.fn()
-
-      const iconOptions = options.map((option) => ({
-        ...option,
-        icon: <IconAnchor data-testid="select-option-icon" />,
-      }))
-
-      wrapper = render(<Select label="Select label" options={iconOptions} />)
-
-      const input = wrapper.getByTestId('react-select-vendor-input')
-      fireEvent.focus(input)
-      fireEvent.keyDown(input, {
-        key: 'ArrowDown',
-        code: 40,
-      })
-    })
-
-    it('should render the icons', () => {
-      expect(wrapper.getAllByTestId('select-option-icon')).toHaveLength(2)
-    })
-  })
-
-  describe('when there is a clear button outside', () => {
-    beforeEach(() => {
-      const SelectWithOutsideClearButton: React.FC = () => {
-        const [value, setValue] = useState<string>()
-        const onChange = (event: any) => setValue(event.target.value)
-        const onClick = () => setValue(null)
-
-        return (
-          <>
-            <Select
-              options={options}
-              value={value}
-              onChange={onChange}
-              label="test"
-            />
-            <Button data-testid="clear-button" onClick={onClick}>
-              Clear
-            </Button>
-          </>
-        )
-      }
-
-      wrapper = render(<SelectWithOutsideClearButton />)
-    })
-
-    describe('and an option is selected', () => {
-      beforeEach(async () => {
-        const input = wrapper.getByTestId('react-select-vendor-input')
-        fireEvent.focus(input)
-        fireEvent.keyDown(input, {
-          key: 'ArrowDown',
-          code: 40,
-        })
-
-        wrapper.getByText('Option 1').click()
-      })
-
-      describe('and the `Clear` button is clicked', () => {
-        beforeEach(() => {
-          wrapper.getByTestId('clear-button').click()
-        })
-
-        it('should clear the select label', () => {
-          return waitFor(() => {
-            expect(
-              wrapper.queryAllByTestId('select-single-value-label')
-            ).toHaveLength(0)
-          })
-        })
-      })
-    })
-  })
-
-  describe('when `isClearable` is `false`', () => {
-    beforeEach(() => {
-      onChangeSpy = jest.fn()
-
-      wrapper = render(
-        <Select isClearable={false} onChange={onChangeSpy} options={options} />
-      )
-    })
-
-    describe('when the select is clicked', () => {
+    describe('when tabbing to the input', () => {
       beforeEach(() => {
-        const input = wrapper.getByTestId('react-select-vendor-input')
-        fireEvent.focus(input)
-        fireEvent.keyDown(input, {
-          key: 'ArrowDown',
-          code: 40,
-        })
+        userEvent.tab()
       })
 
-      describe('when the first option is clicked', () => {
-        beforeEach(() => {
-          wrapper.getByText('Option 1').click()
-        })
-
-        it('should not render a clear button', () => {
-          const control = wrapper.container.children[0].children[1]
-          const indicators = control.children[1]
-          const firstIndicator = indicators.children[0]
-
-          expect(firstIndicator.classList).not.toContain(
-            'rn-select__clear-indicator'
-          )
-        })
+      it('shows the items', () => {
+        expect(wrapper.queryAllByTestId('select-option')).toHaveLength(3)
       })
     })
-  })
 
-  describe('when the options are updated externally', () => {
-    beforeEach(() => {
-      const initialOptions = options
-      const updatedOptions = [
-        {
-          badge: 'Badge 2',
-          label: 'Option 2',
-          value: '2',
-        },
-        {
-          badge: 'Badge 3',
-          label: 'Option 3',
-          value: '3',
-        },
-      ]
-
-      const SelectWithUpdate = () => {
-        const [selectOptions, updateSelectOptions] = useState(initialOptions)
-        const [value, setValue] = useState('2')
-
-        return (
-          <>
-            <Button
-              onClick={() => {
-                updateSelectOptions(updatedOptions)
-                setValue('3')
-              }}
-            >
-              Update
-            </Button>
-            <Select options={selectOptions} value={value} />
-          </>
-        )
-      }
-
-      wrapper = render(<SelectWithUpdate />)
-
-      wrapper.getByText('Update').click()
-    })
-
-    it('should render the select with the selected value', () => {
-      return waitFor(() => {
-        expect(
-          wrapper.getByTestId('select-single-value-label')
-        ).toHaveTextContent('Option 3')
-      })
-    })
-  })
-
-  describe('when an option value is selected that overflows', () => {
-    const label =
-      'This is a really, really long select option label that overflows the container when selected'
-
-    beforeEach(() => {
-      wrapper = render(
-        <Select
-          options={[
-            ...options,
-            {
-              label,
-              value: 'foo',
-            },
-          ]}
-        />
-      )
-
-      const input = wrapper.getByTestId('react-select-vendor-input')
-      fireEvent.focus(input)
-      fireEvent.keyDown(input, {
-        key: 'ArrowDown',
-        code: 40,
-      })
-
-      userEvent.click(wrapper.getByText(label))
-    })
-
-    describe('when the user hovers over the label', () => {
+    describe('when clicking on the input', () => {
       beforeEach(() => {
-        userEvent.hover(wrapper.getByText(label))
+        userEvent.click(wrapper.getByTestId('select-input'))
       })
 
-      it('displays the Tooltip with the label content', () => {
-        expect(wrapper.getByTestId('floating-box')).toBeInTheDocument()
-        expect(wrapper.getByTestId('floating-box-content')).toHaveTextContent(
-          label
-        )
+      it('shows the items', () => {
+        expect(wrapper.queryAllByTestId('select-option')).toHaveLength(3)
+      })
+
+      it('does not display badges', () => {
+        expect(wrapper.queryAllByTestId('select-badge')).toHaveLength(0)
       })
     })
   })
 
-  describe('when the optional `onChange` callback is not provided and the user selects an item', () => {
+  describe('when disabled', () => {
     beforeEach(() => {
       wrapper = render(
-        <Select
-          options={[
-            ...options,
-            {
-              label: 'Foo',
-              value: 'foo',
-            },
-          ]}
-        />
+        <Select isDisabled label="Label">
+          <SelectOption value="one">One</SelectOption>
+          <SelectOption value="two">Two</SelectOption>
+          <SelectOption value="three">Three</SelectOption>
+        </Select>
       )
+    })
 
-      const input = wrapper.getByTestId('react-select-vendor-input')
-      fireEvent.focus(input)
-      fireEvent.keyDown(input, {
-        key: 'ArrowDown',
-        code: 40,
+    describe('when the mouse enters the input', () => {
+      beforeEach(() => {
+        fireEvent.mouseEnter(wrapper.getByTestId('select-input'))
+      })
+
+      it.skip('does not show the arrow button in a hover state', () => {
+        expect(wrapper.getByTestId('select-arrow-button')).toHaveStyleRule(
+          'background-color',
+          'transparent'
+        )
       })
     })
 
-    it('should not attempt to invoke the `onChange` callback', () => {
-      expect(() => userEvent.click(wrapper.getByText('Foo'))).not.toThrow()
+    describe('when tabbing to the input', () => {
+      beforeEach(() => {
+        userEvent.tab()
+      })
+
+      it('does not show the items', () => {
+        expect(wrapper.queryAllByTestId('select-option')).toHaveLength(0)
+      })
+    })
+
+    describe('when clicking on the input', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByTestId('select-input'))
+      })
+
+      it('does not show the items', () => {
+        expect(wrapper.queryAllByTestId('select-option')).toHaveLength(0)
+      })
+    })
+
+    describe('when clicking on the arrow button', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByTestId('select-arrow-button'))
+      })
+
+      it('does not show the items', () => {
+        expect(wrapper.queryAllByTestId('select-option')).toHaveLength(0)
+      })
+    })
+  })
+
+  describe('when invalid', () => {
+    beforeEach(() => {
+      wrapper = render(
+        <Select isInvalid label="Label">
+          <SelectOption value="one">One</SelectOption>
+          <SelectOption value="two">Two</SelectOption>
+          <SelectOption value="three">Three</SelectOption>
+        </Select>
+      )
+    })
+
+    it('displays in an error state', () => {
+      expect(wrapper.getByTestId('select-outer-wrapper')).toHaveStyleRule(
+        'box-shadow',
+        `0 0 0 ${
+          BORDER_WIDTH[COMPONENT_SIZE.FORMS]
+        } ${ColorDanger800.toUpperCase()}`
+      )
+    })
+  })
+
+  describe('when `value` is set to a valid value', () => {
+    beforeEach(() => {
+      wrapper = render(
+        <Select label="Label" value="two">
+          <SelectOption value="one">One</SelectOption>
+          <SelectOption value="two">Two</SelectOption>
+          <SelectOption value="three">Three</SelectOption>
+        </Select>
+      )
+    })
+
+    it('sets the value', () => {
+      expect(wrapper.getByTestId('select-input')).toHaveValue('Two')
+    })
+  })
+
+  describe.each(['invalid', null])('when `value` is set to `%s`', (value) => {
+    beforeEach(() => {
+      wrapper = render(
+        <Select label="Label" value={value}>
+          <SelectOption value="one">One</SelectOption>
+          <SelectOption value="two">Two</SelectOption>
+          <SelectOption value="three">Three</SelectOption>
+        </Select>
+      )
+    })
+
+    it('does not set the value', () => {
+      expect(wrapper.getByTestId('select-input')).toHaveValue('')
+    })
+  })
+
+  describe('when options have icons', () => {
+    beforeEach(() => {
+      wrapper = render(
+        <Select id="select-id" label="Label">
+          <SelectOption
+            icon={<IconAnchor data-testid="select-icon" />}
+            value="one"
+          >
+            One
+          </SelectOption>
+          <SelectOption
+            icon={<IconRemove data-testid="select-icon" />}
+            value="two"
+          >
+            Two
+          </SelectOption>
+          <SelectOption
+            icon={<IconAgriculture data-testid="select-icon" />}
+            value="three"
+          >
+            Three
+          </SelectOption>
+        </Select>
+      )
+    })
+
+    describe('and clicking on the input', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByTestId('select-input'))
+      })
+
+      it('displays 3 icons', () => {
+        expect(wrapper.queryAllByTestId('select-icon')).toHaveLength(3)
+      })
+    })
+  })
+
+  describe('when options have badges', () => {
+    beforeEach(() => {
+      wrapper = render(
+        <Select id="select-id" label="Label">
+          <SelectOption badge={1} value="one">
+            One
+          </SelectOption>
+          <SelectOption badge={2} value="two">
+            Two
+          </SelectOption>
+          <SelectOption badge={3} value="three">
+            Three
+          </SelectOption>
+        </Select>
+      )
+    })
+
+    describe('when clicking on the input', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByTestId('select-input'))
+      })
+
+      it('displays 3 badges', () => {
+        const badges = wrapper.queryAllByTestId('select-badge')
+        expect(badges[0]).toHaveTextContent('1')
+        expect(badges[1]).toHaveTextContent('2')
+        expect(badges[2]).toHaveTextContent('3')
+        expect(badges).toHaveLength(3)
+      })
     })
   })
 })
