+++ packages/react-component-library/src/components/Select/Select.tsx
@@ -1,101 +1,92 @@
 import React from 'react'
-import ReactSelect, { Props as ReactSelectProps } from 'react-select'
+import { useSelect } from 'downshift'
 
-import { DropdownIndicator } from '../Dropdown/DropdownIndicator'
-import { Input } from './Input'
-import { SingleValue } from './SingleValue'
-import { Option, SelectOptionWithBadgeType } from './Option'
-import { ComponentWithClass } from '../../common/ComponentWithClass'
-import { InputValidationProps } from '../../common/InputValidationProps'
-import { StyledMenuList } from './partials/StyledMenuList'
-import { StyledMenu } from './partials/StyledMenu'
-import { StyledValueContainer } from './partials/StyledValueContainer'
-import { StyledIndicatorSeparator } from './partials/StyledIndicatorSeparator'
-import { StyledControl } from './partials/StyledControl'
-import { useSelectedOption } from './useSelectedOption'
+import { getId } from '../../helpers'
+import {
+  initialSelectedItem,
+  itemToString,
+  SelectBaseProps,
+  SelectChildWithStringType,
+  SelectLayout,
+} from '../SelectBase'
 
-export interface SelectProps
-  extends ReactSelectProps<any>,
-    ComponentWithClass,
-    InputValidationProps {
-  /**
-   * Specify whether or not to have a clear button.
-   */
-  isClearable?: boolean
-  /**
-   * Optional text label to display within the component.
-   */
-  label?: string
-  /**
-   * Optional HTML `name` attribute to apply to the component.
-   */
-  name?: string
-  /**
-   * Optional handler invoked when the selected value changes.
-   */
-  onChange?: (event: any) => void
-  /**
-   * Collection of options to display within the Select (JSON).
-   */
-  options: SelectOptionWithBadgeType[]
-  /**
-   * Optional HTML `value` attribute to apply to the component.
-   */
-  value?: string
-}
-
-export const Select: React.FC<SelectProps> = ({
-  isClearable = true,
-  label,
-  name,
+export const Select: React.FC<SelectBaseProps> = ({
+  children,
+  id = getId('select'),
   onChange,
-  options,
-  value,
-  defaultMenuIsOpen,
+  value = null,
   ...rest
 }) => {
-  const { selectedOption, setSelectedValue } = useSelectedOption(options, value)
-
-  const onSelectChange = (option: any) => {
-    const selectedValue =
-      option && option.value !== undefined ? option.value : null
-    setSelectedValue(selectedValue)
-
-    if (onChange) {
-      onChange({
-        target: {
-          name,
-          value: selectedValue,
-        },
-      })
-    }
-  }
+  const {
+    getItemProps,
+    getMenuProps,
+    getToggleButtonProps,
+    highlightedIndex,
+    isOpen,
+    openMenu,
+    reset,
+    selectedItem,
+    toggleMenu,
+  } = useSelect<SelectChildWithStringType>({
+    itemToString,
+    initialSelectedItem: initialSelectedItem(children, value),
+    items: React.Children.toArray(children),
+    onSelectedItemChange: ({ selectedItem: newItem }) => {
+      if (onChange) {
+        onChange(React.isValidElement(newItem) ? newItem.props.value : null)
+      }
+    },
+  })
 
   return (
-    <ReactSelect
-      aria-label={label}
-      classNamePrefix="rn-select"
-      components={{
-        DropdownIndicator,
-        Input,
-        Option,
-        SingleValue,
-        Control: StyledControl,
-        IndicatorSeparator: StyledIndicatorSeparator,
-        Menu: StyledMenu,
-        MenuList: StyledMenuList,
-        ValueContainer: StyledValueContainer,
+    <SelectLayout
+      hasSelectedItem={!!selectedItem}
+      id={id}
+      inputProps={{
+        onFocus: () => {
+          if (!isOpen) {
+            openMenu()
+          }
+        },
+        onMouseDown: (e: React.MouseEvent) => {
+          toggleMenu()
+          e.stopPropagation()
+          e.preventDefault()
+        },
       }}
-      isClearable={isClearable}
-      name={name}
-      onChange={onSelectChange}
-      options={options}
-      placeholder={null}
-      value={selectedOption}
-      autoFocus={defaultMenuIsOpen}
-      defaultMenuIsOpen={defaultMenuIsOpen}
-      {...rest}
-    />
+      isOpen={isOpen}
+      menuProps={getMenuProps()}
+      onClearButtonClick={() => {
+        reset()
+      }}
+      toggleButtonProps={getToggleButtonProps()}
+      value={selectedItem ? itemToString(selectedItem) : ''}
+      {...{
+        readOnly: true,
+        ...rest,
+      }}
+    >
+      {isOpen &&
+        React.Children.map(
+          children,
+          (child: SelectChildWithStringType, index) => {
+            if (!React.isValidElement(child)) {
+              return null
+            }
+
+            return React.cloneElement(child, {
+              ...child.props,
+              ...getItemProps({
+                index,
+                item: child,
+                key: `select-option-${child.props.children}`,
+              }),
+              isHighlighted: highlightedIndex === index,
+              title: child.props.children,
+            })
+          }
+        )}
+    </SelectLayout>
   )
 }
