+++ packages/react-component-library/src/components/Select/index.ts
@@ -1,1 +1,2 @@
 export * from './Select'
+export * from './SelectOption'
