--- packages/react-component-library/src/components/Select/partials/StyledBadge.tsx
@@ -1,10 +1,0 @@
-import { selectors } from '@defencedigital/design-tokens'
-import styled from 'styled-components'
-import { Badge, BadgeProps } from '../../Badge'
-
-const { spacing } = selectors
-
-export const StyledBadge = styled(Badge)<BadgeProps>`
-  transform: translateY(-1px);
-  margin-left: ${spacing('2')};
-`
