--- packages/react-component-library/src/components/Select/partials/StyledInput.tsx
@@ -1,14 +1,0 @@
-import { components } from 'react-select'
-import { selectors } from '@defencedigital/design-tokens'
-import styled from 'styled-components'
-
-const { fontSize, spacing } = selectors
-
-interface StyledInputProps {
-  name: string
-}
-
-export const StyledInput = styled(components.Input)<StyledInputProps>`
-  font-size: ${fontSize('base')};
-  margin: ${spacing('8')} 0 0;
-`
