--- packages/react-component-library/src/components/Select/partials/StyledInputContainer.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledInputContainer = styled.div`
-  display: flex;
-  align-items: center;
-  height: inherit;
-
-  &&& > div {
-    padding: 0;
-    margin: 0;
-  }
-`
