--- packages/react-component-library/src/components/Select/partials/StyledOptionEndAdornment.tsx
@@ -1,12 +1,0 @@
-import { selectors } from '@defencedigital/design-tokens'
-import styled from 'styled-components'
-
-const { spacing } = selectors
-
-export const StyledOptionEndAdornment = styled.span`
-  position: absolute;
-  right: ${spacing('4')};
-  top: 50%;
-  transform: translateY(-50%);
-  height: ${spacing('7')};
-`
