--- packages/react-component-library/src/components/Select/partials/StyledOptionLabel.tsx
@@ -1,9 +1,0 @@
-import { selectors } from '@defencedigital/design-tokens'
-import styled from 'styled-components'
-
-const { fontSize } = selectors
-
-export const StyledOptionLabel = styled.span`
-  font-size: ${fontSize('base')};
-  font-weight: 500;
-`
