--- packages/react-component-library/src/components/Select/partials/StyledSingleValueLabel.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledSingleValueLabel = styled.span`
-  display: inline-block;
-  max-width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  padding: 6px 0;
-`
