+++ packages/react-component-library/src/components/SelectBase/SelectLayout.tsx
@@ -2,8 +2,8 @@ import { UseSelectReturnValue, UseComboboxReturnValue } from 'downshift'
 import React, { useState } from 'react'
 import mergeRefs from 'react-merge-refs'
 
-import { ArrowButton } from '../SelectE/ArrowButton'
-import { ClearButton } from '../SelectE/ClearButton'
+import { ArrowButton } from '../Select/ArrowButton'
+import { ClearButton } from '../Select/ClearButton'
 import { ComponentWithClass } from '../../common/ComponentWithClass'
 import { FloatingBox } from '../../primitives/FloatingBox'
 import { SelectChildWithStringType } from './types'
