--- packages/react-component-library/src/components/SelectE/index.ts
@@ -1,2 +1,0 @@
-export * from './SelectE'
-export * from './SelectEOption'
