+++ packages/react-component-library/src/index.ts
@@ -41,7 +41,6 @@ export * from './components/Radio'
 export * from './components/RangeSlider'
 export * from './components/SectionDivider'
 export * from './components/Select'
-export * from './components/SelectE'
 export * from './components/Switch'
 export * from './components/SwitchE'
 export * from './components/Table'
